#40 - split TeamRequest for Activate and Deactivate Requests
